requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
mysql-connector-python==8.3.0
pydantic==2.6.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
